Year units

era_year objects describe the unit used for a year as its length in days. This value is used in an era definition (era()) to enable conversions between eras that use different units (with yr_transform()).

era_year(label = character(), days = 365.2425)

Arguments

label

Character. Name of the year unit.

days

Numeric. Average length of the year in solar days. Defaults to a Gregorian year (365.2425 days).

Value

S3 vector of class era_year.

See also

Examples

era_year("Julian", 365.25)
#> <era_year[1]>
#> [1] Julian years (365.25 days)